+++ src/main/java/mju/chatuniv/chat/application/dto/gpt/ChatRequest.java
@@ -1,4 +1,4 @@
-package mju.chatuniv.chat.application.dto;
+package mju.chatuniv.chat.application.dto.gpt;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -10,7 +10,6 @@ public class ChatRequest {
 
     public ChatRequest(final String model, final String prompt) {
         this.model = model;
-
         this.messages = new ArrayList<>();
         this.messages.add(new Message("system", "You are a helpful assistant. about 명지대학교 in korea"));
         this.messages.add(new Message("user", prompt));
