+++ src/main/java/mju/chatuniv/global/config/WebMvcConfig.java
@@ -1,6 +1,6 @@
-package mju.chatuniv.config;
+package mju.chatuniv.global.config;
 
-import mju.chatuniv.config.filter.AuthorizationFilter;
+import mju.chatuniv.global.config.filter.AuthorizationFilter;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
